curve out "cut" filter into separated dir
