+++ filter/cut/cut_test.go
@@ -1,4 +1,4 @@
-package filter
+package cut
 
 import (
 	"bytes"
